Update README.md
Added the text "Description: Some goals I have for this class are learning more about analyzing biological data sets, so I can get a better understanding of data analysis for parasitological studies. I am also hoping to gain a proficiency in R, Github, and any other software we plan on using in this class." to the README file.